requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.4.26
charset-normalizer==3.4.2
datasets==3.5.1
dill==0.3.8
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
lxml==5.4.0
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
networkx==3.2.1
numpy==2.0.2
openai==1.77.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
propcache==0.3.1
psycopg2-binary==2.9.10
pyarrow==20.0.0
pydantic==2.11.4
pydantic_core==2.33.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
selenium==4.32.0
sentence-transformers==4.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
websocket-client==1.8.0
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.0
